% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{add_list_to_global}
\alias{add_list_to_global}
\title{add_list_to_global}
\usage{
add_list_to_global(x, only_dfs = FALSE)
}
\arguments{
\item{x}{a named list}

\item{only_dfs}{logical for only including data.frames in output}
}
\value{
environment adds
}
\description{
add_list_to_global
}
