% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{add_to_global}
\alias{add_to_global}
\title{add_to_global}
\usage{
add_to_global(..., only_dfs = FALSE)
}
\arguments{
\item{only_dfs}{logical for only including data.frames in output}
}
\value{
environment adds
}
\description{
add_to_global
}
