% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console.R
\name{bullet_in_console}
\alias{bullet_in_console}
\title{clickable_url_in_console}
\usage{
bullet_in_console(text = "", url = NULL, bullet_type = "i", collape_urls = F)
}
\description{
clickable_url_in_console
}
