% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emails.R
\name{choose_emails_to_delete_from}
\alias{choose_emails_to_delete_from}
\title{choose_emails_to_delete_from}
\usage{
choose_emails_to_delete_from(
  outlook,
  address,
  n = 2000,
  individual_choice = F,
  ask = T
)
}
\description{
choose_emails_to_delete_from
}
