% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emails.R
\name{choose_emails_to_delete_in_bulk}
\alias{choose_emails_to_delete_in_bulk}
\title{choose_emails_to_delete_in_bulk}
\usage{
choose_emails_to_delete_in_bulk(inbox, full_address = T, n = 1000)
}
\description{
choose_emails_to_delete_in_bulk
}
