% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console.R
\name{cli_alert_wrap}
\alias{cli_alert_wrap}
\title{cli_alert_wrap}
\usage{
cli_alert_wrap(
  text = "",
  url = NULL,
  bullet_type = "i",
  collapse = TRUE,
  file = NULL,
  silent = FALSE
)
}
\description{
cli_alert_wrap
}
