% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{combine_files}
\alias{combine_files}
\title{combine_files}
\usage{
combine_files(dir, starts_with, ext = "xlsx")
}
\arguments{
\item{dir}{a dir path}

\item{starts_with}{character for file_name startsWith}

\item{ext}{character file extention}
}
\value{
data.frame
}
\description{
combine_files
}
