% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{default_instrument_remap}
\alias{default_instrument_remap}
\title{default_instrument_remap}
\usage{
default_instrument_remap(instruments_remap, merge_form_name = "merged")
}
\description{
default_instrument_remap
}
