% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{delete_file}
\alias{delete_file}
\title{delete_file}
\usage{
delete_file(path, silent = FALSE)
}
\arguments{
\item{path}{a file path}

\item{silent}{logical for silencing messages}
}
\value{
file deleted and maybe a message
}
\description{
delete_file
}
