% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{form_to_wb}
\alias{form_to_wb}
\title{form_to_wb}
\usage{
form_to_wb(
  form,
  form_name,
  wb = openxlsx::createWorkbook(),
  key_cols = NULL,
  derived_cols = NULL,
  link_col_list = list(),
  str_trunc_length = 32000,
  header_df = NULL,
  tableStyle = "none",
  header_style = default_header_style,
  body_style = default_body_style,
  freeze_header = TRUE,
  pad_rows = 0,
  pad_cols = 0,
  freeze_keys = TRUE
)
}
\description{
form_to_wb
}
