% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{full_file_info}
\alias{full_file_info}
\title{full_file_info}
\usage{
full_file_info(path, showWarnings = TRUE)
}
\arguments{
\item{path}{a file path}

\item{showWarnings}{logical for showing warnings}
}
\value{
file_info data.frame
}
\description{
full_file_info
}
