% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R, R/files.R
\name{list_files_real}
\alias{list_files_real}
\title{list_files_real}
\usage{
list_files_real(path, full.names = TRUE, recursive = FALSE)

list_files_real(path, full.names = TRUE, recursive = FALSE)
}
\description{
list_files_real

list_files_real
}
