% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{list_to_excel}
\alias{list_to_excel}
\title{list_to_wb}
\usage{
list_to_excel(
  list,
  dir,
  file_name = NULL,
  separate = FALSE,
  overwrite = TRUE,
  link_col_list = list(),
  str_trunc_length = 32000,
  header_df_list,
  tableStyle = "none",
  header_style = openxlsx::createStyle(),
  body_style = openxlsx::createStyle(),
  freeze_header = T
)
}
\description{
list_to_wb
}
