% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{list_to_wb}
\alias{list_to_wb}
\title{list_to_wb}
\usage{
list_to_wb(
  list,
  link_col_list = list(),
  str_trunc_length = 32000,
  header_df_list = list(),
  tableStyle = "none",
  header_style = openxlsx::createStyle(),
  body_style = openxlsx::createStyle()
)
}
\description{
list_to_wb
}
