% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{mod_backend_server}
\alias{mod_backend_server}
\title{mod_backend_server}
\usage{
mod_backend_server(id, input, values = NULL)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
