% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{mod_list_server}
\alias{mod_list_server}
\title{mod_backend_server}
\usage{
mod_list_server(id, values)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
