% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{most_recent_file}
\alias{most_recent_file}
\title{most_recent_file}
\usage{
most_recent_file(dir, starts_with, ext = "xlsx")
}
\arguments{
\item{dir}{a dir path}

\item{starts_with}{character for file_name startsWith}

\item{ext}{character file extention}
}
\value{
file_info data.frame
}
\description{
most_recent_file
}
