% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{purify_names_df_list}
\alias{purify_names_df_list}
\title{purify_names_df_list}
\usage{
purify_names_df_list(
  named_df_list,
  instruments_remap,
  metadata,
  merge_list = NULL
)
}
\description{
purify_names_df_list
}
