% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{reassign_variable_in_bulk}
\alias{reassign_variable_in_bulk}
\title{reassign_variable_in_bulk}
\usage{
reassign_variable_in_bulk(DF, old_colname, new_colname, optional_choices)
}
\description{
reassign_variable_in_bulk
}
