% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{recursive_markdown}
\alias{recursive_markdown}
\title{validate_web_link}
\usage{
recursive_markdown(
  DF,
  order_master,
  level = 1,
  display_cols = NULL,
  link_cols = NULL,
  bold_cols = NULL,
  named_cols = NULL,
  header_col = NULL,
  name_master = NULL,
  new_pages_levels = NULL,
  newpage_triggered = FALSE
)
}
\description{
validate_web_link
}
