% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{remap_named_df_list}
\alias{remap_named_df_list}
\title{remap_named_df_list}
\usage{
remap_named_df_list(
  named_df_list,
  instruments_remap,
  merge_form_name = "merged",
  merge_list = NULL
)
}
\description{
remap_named_df_list
}
