% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{unnest_dir}
\alias{unnest_dir}
\title{unnest_dir}
\usage{
unnest_dir(from, to, ignore_folders = NULL)
}
\arguments{
\item{from}{a file path for from}

\item{to}{a file path for to}

\item{ignore_folders}{character string of folders to ignore}
}
\value{
message
}
\description{
unnest_dir
}
