% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{view_file}
\alias{view_file}
\title{view_file}
\usage{
view_file(path, browser = F)
}
\arguments{
\item{browser}{logical for launching from your PC default, not RStudio}
}
\value{
file opens
}
\description{
view_file
}
